@@ -20,10 +20,10 @@
         }
     },
     "require": {
-        "php": "^8.0",
-        "symfony/dependency-injection": "^5.4.21 || ^6.0",
-        "symfony/http-kernel": "^5.4.21 || ^6.0",
-        "symfony/workflow": "^5.4.21 || ^6.0",
+        "php": "^8.2",
+        "symfony/dependency-injection": "^6.4.1",
+        "symfony/http-kernel": "^6.4.1",
+        "symfony/workflow": "^6.4.0",
         "winzou/state-machine": "^0.4",
         "winzou/state-machine-bundle": "^0.6"
     },
@@ -34,7 +34,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-main": "1.13-dev"
+            "dev-main": "2.0-dev"
         },
         "symfony": {
             "require": "^5.4.21"
