@@ -16,7 +16,6 @@ namespace Tests\Sylius\Abstraction\StateMachine\Unit;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use Sylius\Abstraction\StateMachine\Exception\StateMachineExecutionException;
-use Sylius\Abstraction\StateMachine\StateMachineInterface;
 use Sylius\Abstraction\StateMachine\SymfonyWorkflowAdapter;
 use Symfony\Component\Workflow\Exception\InvalidArgumentException;
 use Symfony\Component\Workflow\Registry;
@@ -150,13 +149,126 @@ final class SymfonyWorkflowAdapterTest extends TestCase
         $this->createTestSubject()->getEnabledTransitions($subject, $graphName);
     }
 
-    private function createTestSubject(): StateMachineInterface
+    /**
+     * @dataProvider itReturnsTransitionsToForGivenTransitionProvider
+     */
+    public function testItReturnsTransitionsToForGivenTransition(?string $expectedToStateTransition, string $requestedTransition): void
+    {
+        $subject = new \stdClass();
+        $graphName = 'graph_name';
+
+        $someTransition = $this->createSampleTransition(
+            name: 'some_transition',
+            froms: ['from_1', 'from_2'],
+            tos: ['to_1'],
+        );
+        $anotherTransition = $this->createSampleTransition(
+            name: 'another_transition',
+            froms: ['from_1', 'from_3'],
+            tos: ['to_1', 'to_2'],
+        );
+
+        $workflow = $this->createMock(Workflow::class);
+        $workflow->method('getEnabledTransitions')->with($subject)->willReturn([$someTransition, $anotherTransition]);
+
+        $this->symfonyWorkflowRegistry->method('get')->with($subject, $graphName)->willReturn($workflow);
+
+        $testSubject = $this->createTestSubject();
+
+        $this->assertSame(
+            $expectedToStateTransition,
+            $testSubject->getTransitionToState($subject, $graphName, $requestedTransition),
+        );
+    }
+
+    /**
+     * @return iterable<string, mixed>
+     */
+    public static function itReturnsTransitionsToForGivenTransitionProvider(): iterable
+    {
+        yield 'it returns first transition for to_1' => [
+            'expectedToStateTransition' => 'some_transition',
+            'requestedTransition' => 'to_1',
+        ];
+        yield ' it returns second transition for to_2' => [
+            'expectedToStateTransition' => 'another_transition',
+            'requestedTransition' => 'to_2',
+        ];
+        yield 'it returns null for to_3' => [
+            'expectedToStateTransition' => null,
+            'requestedTransition' => 'to_3',
+        ];
+    }
+
+    /**
+     * @dataProvider itReturnsTransitionsFromForGivenTransitionProvider
+     */
+    public function testItReturnsTransitionsFromForGivenTransition(?string $expectedFromStateTransition, string $requestedTransition): void
+    {
+        $subject = new \stdClass();
+        $graphName = 'graph_name';
+
+        $someTransition = $this->createSampleTransition(
+            name: 'some_transition',
+            froms: ['from_1', 'from_2'],
+            tos: ['to_1'],
+        );
+        $anotherTransition = $this->createSampleTransition(
+            name: 'another_transition',
+            froms: ['from_1', 'from_3'],
+            tos: ['to_1', 'to_2'],
+        );
+
+        $workflow = $this->createMock(Workflow::class);
+        $workflow->method('getEnabledTransitions')->with($subject)->willReturn([$someTransition, $anotherTransition]);
+
+        $this->symfonyWorkflowRegistry->method('get')->with($subject, $graphName)->willReturn($workflow);
+
+        $testSubject = $this->createTestSubject();
+
+        $this->assertSame(
+            $expectedFromStateTransition,
+            $testSubject->getTransitionFromState($subject, $graphName, $requestedTransition),
+        );
+    }
+
+    /**
+     * @return iterable<string, mixed>
+     */
+    public static function itReturnsTransitionsFromForGivenTransitionProvider(): iterable
+    {
+        yield 'it returns first transition for from_1' => [
+            'expectedFromStateTransition' => 'some_transition',
+            'requestedTransition' => 'from_1',
+        ];
+        yield 'it returns first transition for from_2' => [
+            'expectedFromStateTransition' => 'some_transition',
+            'requestedTransition' => 'from_2',
+        ];
+        yield 'it returns second transition for from_3' => [
+            'expectedFromStateTransition' => 'another_transition',
+            'requestedTransition' => 'from_3',
+        ];
+        yield 'it returns null for from_4' => [
+            'expectedFromStateTransition' => null,
+            'requestedTransition' => 'from_4',
+        ];
+    }
+
+    private function createTestSubject(): SymfonyWorkflowAdapter
     {
         return new SymfonyWorkflowAdapter($this->symfonyWorkflowRegistry);
     }
 
-    private function createSampleTransition(): SymfonyWorkflowTransition
+    private function createSampleTransition(mixed ...$arguments): SymfonyWorkflowTransition
     {
-        return new SymfonyWorkflowTransition('transition', ['from'], ['to']);
+        return new SymfonyWorkflowTransition(...array_replace(
+            [
+                'name' => 'transition',
+                'froms' => ['from'],
+                'tos' => ['to'],
+            ],
+            $arguments,
+        ));
     }
 }
