@@ -17,7 +17,7 @@ use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use SM\Factory\FactoryInterface;
 use SM\SMException;
-use SM\StateMachine\StateMachineInterface as WinzouStateMachineInterface;
+use SM\StateMachine\StateMachine as WinzouStateMachine;
 use Sylius\Abstraction\StateMachine\Exception\StateMachineExecutionException;
 use Sylius\Abstraction\StateMachine\StateMachineInterface;
 use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
@@ -27,14 +27,26 @@ final class WinzouStateMachineAdapterTest extends TestCase
     /** @var FactoryInterface&MockObject */
     private FactoryInterface $winzouStateMachineFactory;
 
-    /** @var WinzouStateMachineInterface&MockObject */
-    private WinzouStateMachineInterface $winzouStateMachine;
+    /** @var WinzouStateMachine&MockObject */
+    private WinzouStateMachine $winzouStateMachine;
 
     protected function setUp(): void
     {
         parent::setUp();
 
-        $this->winzouStateMachine = $this->createMock(WinzouStateMachineInterface::class);
+        $this->winzouStateMachine = $this->createMock(WinzouStateMachine::class);
+        $this->setStateMachineConfig($this->winzouStateMachine, [
+            'transitions' => [
+                'transition' => [
+                    'from' => ['from_state'],
+                    'to' => 'to_state',
+                ],
+                'another_transition' => [
+                    'from' => ['another_from_state'],
+                    'to' => 'another_to_state',
+                ],
+            ],
+        ]);
 
         $this->winzouStateMachineFactory = $this->createMock(FactoryInterface::class);
         $this->winzouStateMachineFactory
@@ -67,17 +79,20 @@ final class WinzouStateMachineAdapterTest extends TestCase
 
     public function testItReturnsEnabledTransitions(): void
     {
-        $this->winzouStateMachine->method('getPossibleTransitions')->willReturn(['transition']);
-
         $subject = new \stdClass();
         $graphName = 'graph_name';
+
+        $this->winzouStateMachine->method('can')->willReturnMap([
+            ['transition', true],
+            ['another_transition', false],
+        ]);
 
         $transitions = $this->createTestSubject()->getEnabledTransitions($subject, $graphName);
 
         $this->assertCount(1, $transitions);
         $this->assertSame('transition', $transitions[0]->getName());
-        $this->assertNull($transitions[0]->getFroms());
-        $this->assertNull($transitions[0]->getTos());
+        $this->assertSame(['from_state'], $transitions[0]->getFroms());
+        $this->assertSame(['to_state'], $transitions[0]->getTos());
     }
 
     public function testItConvertsWorkflowExceptionsToCustomOnesOnCan(): void
@@ -118,8 +133,67 @@ final class WinzouStateMachineAdapterTest extends TestCase
         $this->createTestSubject()->getEnabledTransitions($subject, $graphName);
     }
 
+    public function testItReturnsTransitionsToForGivenTransition(): void
+    {
+        $this->setStateMachineConfig($this->winzouStateMachine, [
+            'transitions' => [
+                'transition_to_state' => [
+                    'from' => ['from_state'],
+                    'to' => 'to_state',
+                ],
+            ],
+        ]);
+
+        $this->winzouStateMachine->method('can')->willReturn(true);
+
+        $this->winzouStateMachineFactory = $this->createMock(FactoryInterface::class);
+        $this->winzouStateMachineFactory->method('get')->willReturn($this->winzouStateMachine);
+
+        $stateMachine = $this->createTestSubject();
+
+        $this->assertSame(
+            'transition_to_state',
+            $stateMachine->getTransitionToState(new \stdClass(), 'graph_name', 'to_state'),
+        );
+    }
+
+    public function testItReturnsTransitionsFromForGivenTransition(): void
+    {
+        $this->setStateMachineConfig($this->winzouStateMachine, [
+            'transitions' => [
+                'transition_from_state' => [
+                    'from' => ['from_state'],
+                    'to' => 'to_state',
+                ],
+            ],
+        ]);
+
+        $this->winzouStateMachine->method('can')->willReturn(true);
+
+        $this->winzouStateMachineFactory = $this->createMock(FactoryInterface::class);
+        $this->winzouStateMachineFactory->method('get')->willReturn($this->winzouStateMachine);
+
+        $stateMachine = $this->createTestSubject();
+
+        $this->assertSame(
+            'transition_from_state',
+            $stateMachine->getTransitionFromState(new \stdClass(), 'graph_name', 'from_state'),
+        );
+    }
+
     private function createTestSubject(): StateMachineInterface
     {
         return new WinzouStateMachineAdapter($this->winzouStateMachineFactory);
+    }
+
+    /**
+     * @param array<string, mixed> $config
+     */
+    private function setStateMachineConfig(WinzouStateMachine $stateMachine, array $config): void
+    {
+        $reflection = new \ReflectionClass($stateMachine);
+        $configProperty = $reflection->getProperty('config');
+        $configProperty->setAccessible(true);
+        $configProperty->setValue($stateMachine, $config);
     }
 }
