@@ -15,7 +15,8 @@ namespace Sylius\Bundle\AdminBundle\Event;
 
 use Knp\Menu\FactoryInterface;
 use Knp\Menu\ItemInterface;
-use SM\StateMachine\StateMachineInterface;
+use SM\StateMachine\StateMachineInterface as WinzouStateMachineInterface;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
 use Sylius\Bundle\UiBundle\Menu\Event\MenuBuilderEvent;
 use Sylius\Component\Core\Model\OrderInterface;
 
@@ -25,9 +26,21 @@ class OrderShowMenuBuilderEvent extends MenuBuilderEvent
         FactoryInterface $factory,
         ItemInterface $menu,
         private OrderInterface $order,
-        private StateMachineInterface $stateMachine,
+        private WinzouStateMachineInterface|StateMachineInterface $stateMachine,
     ) {
         parent::__construct($factory, $menu);
+
+        if ($this->stateMachine instanceof WinzouStateMachineInterface) {
+            trigger_deprecation(
+                'sylius/admin-bundle',
+                '1.13',
+                sprintf(
+                    'Passing an instance of "%s" as the fourth argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                    WinzouStateMachineInterface::class,
+                    StateMachineInterface::class,
+                ),
+            );
+        }
     }
 
     public function getOrder(): OrderInterface
@@ -35,7 +48,7 @@ class OrderShowMenuBuilderEvent extends MenuBuilderEvent
         return $this->order;
     }
 
-    public function getStateMachine(): StateMachineInterface
+    public function getStateMachine(): WinzouStateMachineInterface|StateMachineInterface
     {
         return $this->stateMachine;
     }
