@@ -14,6 +14,8 @@ declare(strict_types=1);
 namespace Sylius\Bundle\ApiBundle\CommandHandler\Checkout;
 
 use SM\Factory\FactoryInterface;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
+use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
 use Sylius\Bundle\ApiBundle\Changer\PaymentMethodChangerInterface;
 use Sylius\Bundle\ApiBundle\Command\Checkout\ChoosePaymentMethod;
 use Sylius\Bundle\ApiBundle\Exception\PaymentMethodCannotBeChangedException;
@@ -33,9 +35,20 @@ final class ChoosePaymentMethodHandler implements MessageHandlerInterface
         private OrderRepositoryInterface $orderRepository,
         private PaymentMethodRepositoryInterface $paymentMethodRepository,
         private PaymentRepositoryInterface $paymentRepository,
-        private FactoryInterface $stateMachineFactory,
+        private FactoryInterface|StateMachineInterface $stateMachineFactory,
         private PaymentMethodChangerInterface $paymentMethodChanger,
     ) {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            trigger_deprecation(
+                'sylius/api-bundle',
+                '1.13',
+                sprintf(
+                    'Passing an instance of "%s" as the fourth argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                    FactoryInterface::class,
+                    StateMachineInterface::class,
+                ),
+            );
+        }
     }
 
     public function __invoke(ChoosePaymentMethod $choosePaymentMethod): OrderInterface
@@ -64,21 +77,27 @@ final class ChoosePaymentMethodHandler implements MessageHandlerInterface
         Assert::notNull($payment, 'Can not find payment with given identifier.');
 
         if ($cart->getState() === OrderInterface::STATE_CART) {
-            $stateMachine = $this->stateMachineFactory->get($cart, OrderCheckoutTransitions::GRAPH);
-
+            $stateMachine = $this->getStateMachine();
             Assert::true(
-                $stateMachine->can(
-                    OrderCheckoutTransitions::TRANSITION_SELECT_PAYMENT,
-                ),
+                $stateMachine->can($cart, OrderCheckoutTransitions::GRAPH, OrderCheckoutTransitions::TRANSITION_SELECT_PAYMENT),
                 'Order cannot have payment method assigned.',
             );
 
             $payment->setMethod($paymentMethod);
-            $stateMachine->apply(OrderCheckoutTransitions::TRANSITION_SELECT_PAYMENT);
+            $stateMachine->apply($cart, OrderCheckoutTransitions::GRAPH, OrderCheckoutTransitions::TRANSITION_SELECT_PAYMENT);
 
             return $cart;
         }
 
         throw new PaymentMethodCannotBeChangedException();
+    }
+
+    private function getStateMachine(): StateMachineInterface
+    {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            return new WinzouStateMachineAdapter($this->stateMachineFactory);
+        }
+
+        return $this->stateMachineFactory;
     }
 }
