@@ -14,7 +14,9 @@ declare(strict_types=1);
 namespace Sylius\Bundle\ApiBundle\CommandHandler\Checkout;
 
 use SM\Factory\FactoryInterface;
-use SM\SMException;
+use Sylius\Abstraction\StateMachine\Exception\StateMachineExecutionException;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
+use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
 use Sylius\Bundle\ApiBundle\Command\Cart\InformAboutCartRecalculation;
 use Sylius\Bundle\ApiBundle\Command\Checkout\CompleteOrder;
 use Sylius\Bundle\ApiBundle\CommandHandler\Checkout\Exception\OrderTotalHasChangedException;
@@ -33,15 +35,26 @@ final class CompleteOrderHandler implements MessageHandlerInterface
 {
     public function __construct(
         private OrderRepositoryInterface $orderRepository,
-        private FactoryInterface $stateMachineFactory,
+        private FactoryInterface|StateMachineInterface $stateMachineFactory,
         private MessageBusInterface $commandBus,
         private MessageBusInterface $eventBus,
         private OrderPromotionsIntegrityCheckerInterface $orderPromotionsIntegrityChecker,
     ) {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            trigger_deprecation(
+                'sylius/api-bundle',
+                '1.13',
+                sprintf(
+                    'Passing an instance of "%s" as the second argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                    FactoryInterface::class,
+                    StateMachineInterface::class,
+                ),
+            );
+        }
     }
 
     /**
-     * @throws SMException
+     * @throws StateMachineExecutionException
      * @throws OrderTotalHasChangedException
      */
     public function __invoke(CompleteOrder $completeOrder): OrderInterface
@@ -73,17 +86,25 @@ final class CompleteOrderHandler implements MessageHandlerInterface
             throw new OrderTotalHasChangedException();
         }
 
-        $stateMachine = $this->stateMachineFactory->get($cart, OrderCheckoutTransitions::GRAPH);
-
+        $stateMachine = $this->getStateMachine();
         Assert::true(
-            $stateMachine->can(OrderCheckoutTransitions::TRANSITION_COMPLETE),
+            $stateMachine->can($cart, OrderCheckoutTransitions::GRAPH, OrderCheckoutTransitions::TRANSITION_COMPLETE),
             sprintf('Order with %s token cannot be completed.', $orderTokenValue),
         );
 
-        $stateMachine->apply(OrderCheckoutTransitions::TRANSITION_COMPLETE);
+        $stateMachine->apply($cart, OrderCheckoutTransitions::GRAPH, OrderCheckoutTransitions::TRANSITION_COMPLETE);
 
         $this->eventBus->dispatch(new OrderCompleted($cart->getTokenValue()), [new DispatchAfterCurrentBusStamp()]);
 
         return $cart;
+    }
+
+    private function getStateMachine(): StateMachineInterface
+    {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            return new WinzouStateMachineAdapter($this->stateMachineFactory);
+        }
+
+        return $this->stateMachineFactory;
     }
 }
