@@ -14,6 +14,8 @@ declare(strict_types=1);
 namespace Sylius\Bundle\ApiBundle\CommandHandler\Checkout;
 
 use SM\Factory\FactoryInterface;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
+use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
 use Sylius\Bundle\ApiBundle\Command\Checkout\SendShipmentConfirmationEmail;
 use Sylius\Bundle\ApiBundle\Command\Checkout\ShipShipment;
 use Sylius\Component\Core\Model\ShipmentInterface;
@@ -29,9 +31,20 @@ final class ShipShipmentHandler implements MessageHandlerInterface
 {
     public function __construct(
         private ShipmentRepositoryInterface $shipmentRepository,
-        private FactoryInterface $stateMachineFactory,
+        private FactoryInterface|StateMachineInterface $stateMachineFactory,
         private MessageBusInterface $eventBus,
     ) {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            trigger_deprecation(
+                'sylius/api-bundle',
+                '1.13',
+                sprintf(
+                    'Passing an instance of "%s" as the second argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                    FactoryInterface::class,
+                    StateMachineInterface::class,
+                ),
+            );
+        }
     }
 
     public function __invoke(ShipShipment $shipShipment): ShipmentInterface
@@ -45,17 +58,25 @@ final class ShipShipmentHandler implements MessageHandlerInterface
             $shipment->setTracking($shipShipment->trackingCode);
         }
 
-        $stateMachine = $this->stateMachineFactory->get($shipment, ShipmentTransitions::GRAPH);
-
+        $stateMachine = $this->getStateMachine();
         Assert::true(
-            $stateMachine->can(ShipmentTransitions::TRANSITION_SHIP),
+            $stateMachine->can($shipment, ShipmentTransitions::GRAPH, ShipmentTransitions::TRANSITION_SHIP),
             'This shipment cannot be completed.',
         );
 
-        $stateMachine->apply(ShipmentTransitions::TRANSITION_SHIP);
+        $stateMachine->apply($shipment, ShipmentTransitions::GRAPH, ShipmentTransitions::TRANSITION_SHIP);
 
         $this->eventBus->dispatch(new SendShipmentConfirmationEmail($shipShipment->shipmentId), [new DispatchAfterCurrentBusStamp()]);
 
         return $shipment;
+    }
+
+    private function getStateMachine(): StateMachineInterface
+    {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            return new WinzouStateMachineAdapter($this->stateMachineFactory);
+        }
+
+        return $this->stateMachineFactory;
     }
 }
