@@ -17,17 +17,19 @@ use ApiPlatform\OpenApi\Model\Operation;
 use ApiPlatform\OpenApi\Model\Parameter;
 use ApiPlatform\OpenApi\Model\PathItem;
 use ApiPlatform\OpenApi\OpenApi;
-use Sylius\Calendar\Provider\DateTimeProviderInterface;
+use Symfony\Component\Clock\ClockInterface;
 use Symfony\Component\HttpFoundation\Response;
 
 /** @experimental */
-final class StatisticsModifier implements DocumentationModifierInterface
+final class StatisticsDocumentationModifier implements DocumentationModifierInterface
 {
     private const PATH = '/admin/statistics';
 
+    /** @param array<string, array<string, string>> $intervalsMap */
     public function __construct(
         private string $apiRoute,
-        private DateTimeProviderInterface $dateTimeProvider,
+        private ClockInterface $clock,
+        private array $intervalsMap,
     ) {
     }
 
@@ -37,7 +39,7 @@ final class StatisticsModifier implements DocumentationModifierInterface
         $schemas['Statistics'] = [
             'type' => 'object',
             'properties' => [
-                'salesPerPeriod' => [
+                'sales' => [
                     'type' => 'array',
                     'items' => [
                         'type' => 'object',
@@ -45,34 +47,38 @@ final class StatisticsModifier implements DocumentationModifierInterface
                             'period' => [
                                 'type' => 'string',
                                 'format' => 'date-time',
+                                'example' => '1999-12',
                             ],
                             'total' => [
                                 'type' => 'integer',
-                                'example' => '1000',
+                                'example' => 1000,
                             ],
                         ],
                     ],
-                    'minItems' => 12,
                 ],
-                'newCustomersCount' => [
-                    'type' => 'integer',
-                    'example' => '10',
-                ],
-                'paidOrdersCount' => [
-                    'type' => 'integer',
-                    'example' => '12',
-                ],
-                'averageOrderValue' => [
-                    'type' => 'integer',
-                    'example' => '1000',
-                ],
-                'totalSales' => [
-                    'type' => 'integer',
-                    'example' => '12000',
-                ],
-                'intervalType' => [
-                    'type' => 'string',
-                    'enum' => ['month', 'year', 'week', 'day'],
+                'businessActivitySummary' => [
+                    'type' => 'array',
+                    'items' => [
+                        'type' => 'object',
+                        'properties' => [
+                            'totalSales' => [
+                                'type' => 'integer',
+                                'example' => 100000,
+                            ],
+                            'paidOrdersCount' => [
+                                'type' => 'integer',
+                                'example' => 12,
+                            ],
+                            'newCustomersCount' => [
+                                'type' => 'integer',
+                                'example' => 7,
+                            ],
+                            'averageOrderValue' => [
+                                'type' => 'integer',
+                                'example' => 2500,
+                            ],
+                        ],
+                    ],
                 ],
             ],
         ];
@@ -135,18 +141,19 @@ final class StatisticsModifier implements DocumentationModifierInterface
             schema: [
                 'type' => 'string',
                 'format' => 'date-time',
-                'default' => $this->dateTimeProvider->now()->format('Y-01-01\T00:00:00'),
+                'default' => $this->clock->now()->format('Y-01-01\T00:00:00'),
             ],
         );
 
-        $dateInterval = new Parameter(
-            name: 'dateInterval',
+        $interval = new Parameter(
+            name: 'interval',
             in: 'query',
-            description: 'Date interval for statistics',
+            description: 'Interval type for statistics',
             required: true,
             schema: [
                 'type' => 'string',
-                'default' => 'P1M',
+                'default' => 'month',
+                'enum' => array_keys($this->intervalsMap),
             ],
         );
 
@@ -158,10 +165,10 @@ final class StatisticsModifier implements DocumentationModifierInterface
             schema: [
                 'type' => 'string',
                 'format' => 'date-time',
-                'default' => $this->dateTimeProvider->now()->format('Y-12-31\T23:59:59'),
+                'default' => $this->clock->now()->format('Y-12-31\T23:59:59'),
             ],
         );
 
-        return [$channelCode, $startDate, $dateInterval, $endDate];
+        return [$channelCode, $startDate, $interval, $endDate];
     }
 }
