@@ -16,7 +16,8 @@ namespace spec\Sylius\Bundle\ApiBundle\CommandHandler\Checkout;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 use SM\Factory\FactoryInterface;
-use SM\StateMachine\StateMachineInterface;
+use SM\StateMachine\StateMachineInterface as WinzouStateMachineInterface;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
 use Sylius\Bundle\ApiBundle\Changer\PaymentMethodChangerInterface;
 use Sylius\Bundle\ApiBundle\Command\Checkout\ChoosePaymentMethod;
 use Sylius\Bundle\ApiBundle\Exception\PaymentMethodCannotBeChangedException;
@@ -55,7 +56,7 @@ final class ChoosePaymentMethodHandlerSpec extends ObjectBehavior
         OrderInterface $cart,
         PaymentInterface $payment,
         PaymentMethodInterface $paymentMethod,
-        StateMachineInterface $stateMachine,
+        WinzouStateMachineInterface $stateMachine,
     ): void {
         $choosePaymentMethod = new ChoosePaymentMethod('CASH_ON_DELIVERY_METHOD');
         $choosePaymentMethod->setOrderTokenValue('ORDERTOKEN');
@@ -78,6 +79,41 @@ final class ChoosePaymentMethodHandlerSpec extends ObjectBehavior
         $payment->setMethod($paymentMethod)->shouldBeCalled();
 
         $stateMachine->apply('select_payment')->shouldBeCalled();
+
+        $this($choosePaymentMethod)->shouldReturn($cart);
+    }
+
+    function it_uses_the_new_state_machine_abstraction_if_passed(
+        OrderRepositoryInterface $orderRepository,
+        PaymentMethodRepositoryInterface $paymentMethodRepository,
+        PaymentRepositoryInterface $paymentRepository,
+        StateMachineInterface $stateMachine,
+        PaymentMethodChangerInterface $paymentMethodChanger,
+        OrderInterface $cart,
+        PaymentInterface $payment,
+        PaymentMethodInterface $paymentMethod,
+    ): void {
+        $this->beConstructedWith($orderRepository, $paymentMethodRepository, $paymentRepository, $stateMachine, $paymentMethodChanger);
+
+        $choosePaymentMethod = new ChoosePaymentMethod('CASH_ON_DELIVERY_METHOD');
+        $choosePaymentMethod->setOrderTokenValue('ORDERTOKEN');
+        $choosePaymentMethod->setSubresourceId('123');
+
+        $orderRepository->findOneBy(['tokenValue' => 'ORDERTOKEN'])->willReturn($cart);
+        $cart->getCheckoutState()->willReturn(OrderCheckoutStates::STATE_SHIPPING_SELECTED);
+
+        $stateMachine->can($cart, OrderCheckoutTransitions::GRAPH, 'select_payment')->willReturn(true);
+        $stateMachine->apply($cart, OrderCheckoutTransitions::GRAPH, 'select_payment')->shouldBeCalled();
+
+        $paymentMethodRepository->findOneBy(['code' => 'CASH_ON_DELIVERY_METHOD'])->willReturn($paymentMethod);
+
+        $cart->getId()->willReturn('111');
+
+        $paymentRepository->findOneByOrderId('123', '111')->willReturn($payment);
+
+        $cart->getState()->willReturn(OrderInterface::STATE_CART);
+
+        $payment->setMethod($paymentMethod)->shouldBeCalled();
 
         $this($choosePaymentMethod)->shouldReturn($cart);
     }
@@ -105,7 +141,7 @@ final class ChoosePaymentMethodHandlerSpec extends ObjectBehavior
         PaymentMethodRepositoryInterface $paymentMethodRepository,
         FactoryInterface $stateMachineFactory,
         OrderInterface $cart,
-        StateMachineInterface $stateMachine,
+        WinzouStateMachineInterface $stateMachine,
         PaymentInterface $payment,
     ): void {
         $choosePaymentMethod = new ChoosePaymentMethod('CASH_ON_DELIVERY_METHOD');
@@ -134,7 +170,7 @@ final class ChoosePaymentMethodHandlerSpec extends ObjectBehavior
         PaymentMethodRepositoryInterface $paymentMethodRepository,
         FactoryInterface $stateMachineFactory,
         OrderInterface $cart,
-        StateMachineInterface $stateMachine,
+        WinzouStateMachineInterface $stateMachine,
         PaymentInterface $payment,
     ): void {
         $choosePaymentMethod = new ChoosePaymentMethod('CASH_ON_DELIVERY_METHOD');
@@ -165,7 +201,7 @@ final class ChoosePaymentMethodHandlerSpec extends ObjectBehavior
         FactoryInterface $stateMachineFactory,
         OrderInterface $cart,
         PaymentMethodInterface $paymentMethod,
-        StateMachineInterface $stateMachine,
+        WinzouStateMachineInterface $stateMachine,
     ): void {
         $choosePaymentMethod = new ChoosePaymentMethod('CASH_ON_DELIVERY_METHOD');
         $choosePaymentMethod->setOrderTokenValue('ORDERTOKEN');
