@@ -17,6 +17,8 @@ use Doctrine\Persistence\ObjectManager;
 use Faker\Factory;
 use Faker\Generator;
 use SM\Factory\FactoryInterface as StateMachineFactoryInterface;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
+use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
 use Sylius\Bundle\CoreBundle\Fixture\OptionsResolver\LazyOption;
 use Sylius\Component\Addressing\Model\CountryInterface;
 use Sylius\Component\Core\Checker\OrderPaymentMethodSelectionRequirementCheckerInterface;
@@ -61,13 +63,25 @@ class OrderExampleFactory extends AbstractExampleFactory implements ExampleFacto
         protected PaymentMethodRepositoryInterface $paymentMethodRepository,
         protected ShippingMethodRepositoryInterface $shippingMethodRepository,
         protected FactoryInterface $addressFactory,
-        protected StateMachineFactoryInterface $stateMachineFactory,
+        protected StateMachineFactoryInterface|StateMachineInterface $stateMachineFactory,
         protected OrderShippingMethodSelectionRequirementCheckerInterface $orderShippingMethodSelectionRequirementChecker,
         protected OrderPaymentMethodSelectionRequirementCheckerInterface $orderPaymentMethodSelectionRequirementChecker,
     ) {
         $this->optionsResolver = new OptionsResolver();
         $this->faker = Factory::create();
         $this->configureOptions($this->optionsResolver);
+
+        if ($this->stateMachineFactory instanceof StateMachineFactoryInterface) {
+            trigger_deprecation(
+                'sylius/core-bundle',
+                '1.13',
+                sprintf(
+                    'Passing an instance of "%s" as the twelfth argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                    StateMachineFactoryInterface::class,
+                    StateMachineInterface::class,
+                ),
+            );
+        }
     }
 
     public function create(array $options = []): OrderInterface
@@ -251,7 +265,7 @@ class OrderExampleFactory extends AbstractExampleFactory implements ExampleFacto
 
     protected function applyCheckoutStateTransition(OrderInterface $order, string $transition): void
     {
-        $this->stateMachineFactory->get($order, OrderCheckoutTransitions::GRAPH)->apply($transition);
+        $this->getStateMachine()->apply($order, OrderCheckoutTransitions::GRAPH, $transition);
     }
 
     protected function generateInvalidSkipMessage(string $type, string $channelCode): string
@@ -281,21 +295,32 @@ class OrderExampleFactory extends AbstractExampleFactory implements ExampleFacto
 
     protected function completePayments(OrderInterface $order): void
     {
+        $stateMachine = $this->getStateMachine();
+
         foreach ($order->getPayments() as $payment) {
-            $stateMachine = $this->stateMachineFactory->get($payment, PaymentTransitions::GRAPH);
-            if ($stateMachine->can(PaymentTransitions::TRANSITION_COMPLETE)) {
-                $stateMachine->apply(PaymentTransitions::TRANSITION_COMPLETE);
+            if ($stateMachine->can($payment, PaymentTransitions::GRAPH, PaymentTransitions::TRANSITION_COMPLETE)) {
+                $stateMachine->apply($payment, PaymentTransitions::GRAPH, PaymentTransitions::TRANSITION_COMPLETE);
             }
         }
     }
 
     protected function completeShipments(OrderInterface $order): void
     {
+        $stateMachine = $this->getStateMachine();
+
         foreach ($order->getShipments() as $shipment) {
-            $stateMachine = $this->stateMachineFactory->get($shipment, ShipmentTransitions::GRAPH);
-            if ($stateMachine->can(ShipmentTransitions::TRANSITION_SHIP)) {
-                $stateMachine->apply(ShipmentTransitions::TRANSITION_SHIP);
+            if ($stateMachine->can($shipment, ShipmentTransitions::GRAPH, ShipmentTransitions::TRANSITION_SHIP)) {
+                $stateMachine->apply($shipment, ShipmentTransitions::GRAPH, ShipmentTransitions::TRANSITION_SHIP);
             }
         }
+    }
+
+    private function getStateMachine(): StateMachineInterface
+    {
+        if ($this->stateMachineFactory instanceof StateMachineFactoryInterface) {
+            return new WinzouStateMachineAdapter($this->stateMachineFactory);
+        }
+
+        return $this->stateMachineFactory;
     }
 }
