@@ -17,6 +17,8 @@ use Doctrine\Persistence\ObjectManager;
 use Faker\Factory;
 use Faker\Generator;
 use SM\Factory\FactoryInterface as StateMachineFactoryInterface;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
+use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
 use Sylius\Bundle\FixturesBundle\Fixture\AbstractFixture;
 use Sylius\Component\Core\Model\PaymentInterface;
 use Sylius\Component\Core\Repository\PaymentRepositoryInterface;
@@ -35,13 +37,25 @@ class PaymentFixture extends AbstractFixture
      */
     public function __construct(
         private PaymentRepositoryInterface $paymentRepository,
-        private StateMachineFactoryInterface $stateMachineFactory,
+        private StateMachineFactoryInterface|StateMachineInterface $stateMachineFactory,
         private ObjectManager $paymentManager,
     ) {
         $this->faker = Factory::create();
 
         $this->optionsResolver = new OptionsResolver();
         $this->configureOptions($this->optionsResolver);
+
+        if ($this->stateMachineFactory instanceof StateMachineFactoryInterface) {
+            trigger_deprecation(
+                'sylius/core-bundle',
+                '1.13',
+                sprintf(
+                    'Passing an instance of "%s" as the second argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                    StateMachineFactoryInterface::class,
+                    StateMachineInterface::class,
+                ),
+            );
+        }
     }
 
     public function getName(): string
@@ -78,11 +92,7 @@ class PaymentFixture extends AbstractFixture
 
     private function completePayment(PaymentInterface $payment): void
     {
-        $this
-            ->stateMachineFactory
-            ->get($payment, PaymentTransitions::GRAPH)
-            ->apply(PaymentTransitions::TRANSITION_COMPLETE)
-        ;
+        $this->getStateMachine()->apply($payment, PaymentTransitions::GRAPH, PaymentTransitions::TRANSITION_COMPLETE);
 
         $this->paymentManager->persist($payment);
     }
@@ -93,5 +103,14 @@ class PaymentFixture extends AbstractFixture
             ->setDefault('percentage_completed', 0)
             ->setAllowedTypes('percentage_completed', 'int')
         ;
+    }
+
+    private function getStateMachine(): StateMachineInterface
+    {
+        if ($this->stateMachineFactory instanceof StateMachineFactoryInterface) {
+            return new WinzouStateMachineAdapter($this->stateMachineFactory);
+        }
+
+        return $this->stateMachineFactory;
     }
 }
