@@ -13,8 +13,11 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\CoreBundle\Validator\Constraints;
 
+use Sylius\Component\Channel\Model\ChannelInterface as BaseChannelInterface;
+use Sylius\Component\Channel\Model\ChannelsAwareInterface;
 use Sylius\Component\Channel\Repository\ChannelRepositoryInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
+use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints\Collection;
 use Symfony\Component\Validator\ConstraintValidator;
@@ -24,8 +27,10 @@ use Symfony\Component\Validator\Exception\UnexpectedValueException;
 final class ChannelCodeCollectionValidator extends ConstraintValidator
 {
     /** @param ChannelRepositoryInterface<ChannelInterface> $channelRepository */
-    public function __construct(private ChannelRepositoryInterface $channelRepository)
-    {
+    public function __construct(
+        private ChannelRepositoryInterface $channelRepository,
+        private PropertyAccessorInterface $propertyAccessor,
+    ) {
     }
 
     public function validate(mixed $value, Constraint $constraint): void
@@ -38,9 +43,43 @@ final class ChannelCodeCollectionValidator extends ConstraintValidator
             throw new UnexpectedValueException($value, 'array');
         }
 
+        if ($constraint->validateAgainstAllChannels) {
+            $this->validateInChannelCollection($value, $this->channelRepository->findAll(), $constraint);
+
+            return;
+        }
+
+        $object = $this->context->getObject();
+
+        if (null !== $constraint->channelAwarePropertyPath) {
+            $object = $this->propertyAccessor->getValue($object, $constraint->channelAwarePropertyPath);
+        }
+
+        if (!$object instanceof ChannelsAwareInterface) {
+            throw new \LogicException(sprintf(
+                'The validated root needs to implement the %s interface when option `validateAgainstAllChannels` is set to false.',
+                ChannelsAwareInterface::class,
+            ));
+        }
+
+        $this->validateInChannelCollection($value, $object->getChannels()->toArray(), $constraint);
+    }
+
+    /**
+     * @param array<array-key, array<array-key, mixed>> $value
+     * @param array<BaseChannelInterface> $channels
+     */
+    private function validateInChannelCollection(
+        array $value,
+        array $channels,
+        ChannelCodeCollection $constraint,
+    ): void {
         $fields = [];
-        foreach ($this->channelRepository->findAll() as $channel) {
+        foreach ($channels as $channel) {
             $fields[$channel->getCode()] = $constraint->constraints;
+        }
+        if ([] === $fields) {
+            return;
         }
 
         $collection = new Collection(
@@ -52,6 +91,7 @@ final class ChannelCodeCollectionValidator extends ConstraintValidator
             $constraint->extraFieldsMessage,
             $constraint->missingFieldsMessage,
         );
+
         $validator = $this->context->getValidator()->inContext($this->context);
         $validator->validate($value, $collection, $constraint->groups);
     }
