@@ -13,10 +13,14 @@ declare(strict_types=1);
 
 namespace spec\Sylius\Bundle\CoreBundle\Validator\Constraints;
 
+use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
 use Sylius\Bundle\CoreBundle\Validator\Constraints\ChannelCodeCollection;
+use Sylius\Component\Channel\Model\ChannelsAwareInterface;
 use Sylius\Component\Channel\Repository\ChannelRepositoryInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
+use Symfony\Component\Form\Form;
+use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\Constraints\Collection;
 use Symfony\Component\Validator\Constraints\NotBlank;
@@ -29,9 +33,9 @@ use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 final class ChannelCodeCollectionValidatorSpec extends ObjectBehavior
 {
-    function let(ChannelRepositoryInterface $channelRepository, ExecutionContextInterface $context): void
+    function let(ChannelRepositoryInterface $channelRepository, PropertyAccessorInterface $propertyAccessor, ExecutionContextInterface $context): void
     {
-        $this->beConstructedWith($channelRepository);
+        $this->beConstructedWith($channelRepository, $propertyAccessor);
 
         $this->initialize($context);
     }
@@ -45,7 +49,7 @@ final class ChannelCodeCollectionValidatorSpec extends ObjectBehavior
         ;
     }
 
-    function it_throws_an_exception_if_value_is_not_an_instance_of_array(): void
+    function it_throws_an_exception_if_value_is_not_an_array(): void
     {
         $this
             ->shouldThrow(UnexpectedValueException::class)
@@ -53,7 +57,131 @@ final class ChannelCodeCollectionValidatorSpec extends ObjectBehavior
         ;
     }
 
-    function it_calls_a_validate_collection_for_channels(
+    function it_throws_exception_when_validating_using_local_channels_and_object_does_not_implement_channels_aware_interface(
+        ExecutionContextInterface $context,
+    ): void {
+        $context->getObject()->willReturn(new \stdClass());
+
+        $this
+            ->shouldThrow(\LogicException::class)
+            ->during('validate', [[], new ChannelCodeCollection(['validateAgainstAllChannels' => false, 'channelAwarePropertyPath' => 'promotion'])])
+        ;
+    }
+
+    function it_retrieves_an_object_from_value_and_validates_collections_for_local_channels(
+        ChannelRepositoryInterface $channelRepository,
+        ChannelInterface $channelMobile,
+        ChannelInterface $channelWeb,
+        ChannelsAwareInterface $channelsAware,
+        ContextualValidatorInterface $contextualValidator,
+        ExecutionContextInterface $context,
+        Form $form,
+        PropertyAccessorInterface $propertyAccessor,
+        ValidatorInterface $validator,
+    ): void {
+        $channelWeb->getCode()->willReturn('WEB');
+        $channelMobile->getCode()->willReturn('MOBILE');
+
+        $channelsAware->getChannels()->willReturn(new ArrayCollection([
+            $channelMobile->getWrappedObject(),
+            $channelWeb->getWrappedObject(),
+        ]));
+
+        $context->getObject()->willReturn($form);
+        $propertyAccessor->getValue($form, 'shippingMethod')->willReturn($channelsAware);
+
+        $constraints = [new NotBlank(), new Type('numeric')];
+        $groups = ['Default', 'test_group'];
+        $value = ['one', 'two'];
+
+        $collection = new Collection(
+            [
+                'WEB' => $constraints,
+                'MOBILE' => $constraints,
+            ],
+            $groups,
+        );
+
+        $channelRepository->findAll()->shouldNotBeCalled();
+
+        $context->getValidator()->willReturn($validator);
+        $validator->inContext($context)->willReturn($contextualValidator);
+
+        $contextualValidator->validate($value, $collection, $groups)->willReturn($contextualValidator)->shouldBeCalled();
+
+        $this->validate($value, new ChannelCodeCollection([
+            'constraints' => $constraints,
+            'groups' => $groups,
+            'channelAwarePropertyPath' => 'shippingMethod',
+        ]));
+    }
+
+    function it_validates_collections_for_local_channels(
+        ChannelRepositoryInterface $channelRepository,
+        ExecutionContextInterface $context,
+        ChannelInterface $channelWeb,
+        ChannelInterface $channelMobile,
+        PropertyAccessorInterface $propertyAccessor,
+        ValidatorInterface $validator,
+        ContextualValidatorInterface $contextualValidator,
+        ChannelsAwareInterface $channelsAware,
+    ): void {
+        $channelWeb->getCode()->willReturn('WEB');
+        $channelMobile->getCode()->willReturn('MOBILE');
+
+        $channelsAware->getChannels()->willReturn(new ArrayCollection([
+            $channelMobile->getWrappedObject(),
+            $channelWeb->getWrappedObject(),
+        ]));
+
+        $context->getObject()->willReturn($channelsAware);
+        $propertyAccessor->getValue($channelsAware, 'promotion')->willReturn($channelsAware);
+
+        $constraints = [new NotBlank(), new Type('numeric')];
+        $groups = ['Default', 'test_group'];
+        $value = ['one', 'two'];
+
+        $collection = new Collection(
+            [
+                'WEB' => $constraints,
+                'MOBILE' => $constraints,
+            ],
+            $groups,
+        );
+
+        $channelRepository->findAll()->shouldNotBeCalled();
+
+        $context->getValidator()->willReturn($validator);
+        $validator->inContext($context)->willReturn($contextualValidator);
+
+        $contextualValidator->validate($value, $collection, $groups)->willReturn($contextualValidator)->shouldBeCalled();
+
+        $this->validate($value, new ChannelCodeCollection([
+            'constraints' => $constraints,
+            'groups' => $groups,
+            'channelAwarePropertyPath' => 'promotion',
+        ]));
+    }
+
+    function it_does_nothing_when_local_collection_if_channels_is_empty(
+        ChannelRepositoryInterface $channelRepository,
+        ExecutionContextInterface $context,
+        ChannelsAwareInterface $channelsAware,
+        PropertyAccessorInterface $propertyAccessor,
+    ): void {
+        $channelsAware->getChannels()->willReturn(new ArrayCollection());
+
+        $context->getObject()->willReturn($channelsAware);
+        $propertyAccessor->getValue($channelsAware, 'promotion')->willReturn($channelsAware);
+
+        $channelRepository->findAll()->shouldNotBeCalled();
+
+        $context->getValidator()->shouldNotBeCalled();
+
+        $this->validate([], new ChannelCodeCollection(['validateAgainstAllChannels' => false, 'channelAwarePropertyPath' => 'promotion']));
+    }
+
+    function it_validates_collections_for_all_channels(
         ChannelRepositoryInterface $channelRepository,
         ExecutionContextInterface $context,
         ChannelInterface $channelWeb,
@@ -82,6 +210,10 @@ final class ChannelCodeCollectionValidatorSpec extends ObjectBehavior
 
         $contextualValidator->validate($value, $collection, $groups)->willReturn($contextualValidator)->shouldBeCalled();
 
-        $this->validate($value, new ChannelCodeCollection(['constraints' => $constraints, 'groups' => $groups]));
+        $this->validate($value, new ChannelCodeCollection([
+            'constraints' => $constraints,
+            'groups' => $groups,
+            'validateAgainstAllChannels' => true,
+        ]));
     }
 }
