@@ -19,16 +19,25 @@ use Payum\Core\Request\Generic;
 use Payum\Core\Request\GetStatusInterface;
 use Payum\Core\Request\Notify;
 use SM\Factory\FactoryInterface;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
+use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
 use Sylius\Bundle\PayumBundle\Request\GetStatus;
 use Sylius\Component\Payment\Model\PaymentInterface;
 use Sylius\Component\Payment\PaymentTransitions;
-use Sylius\Component\Resource\StateMachine\StateMachineInterface;
-use Webmozart\Assert\Assert;
 
 final class UpdatePaymentStateExtension implements ExtensionInterface
 {
-    public function __construct(private FactoryInterface $factory)
+    public function __construct(private FactoryInterface|StateMachineInterface $factory)
     {
+        trigger_deprecation(
+            'sylius/payum-bundle',
+            '1.13',
+            sprintf(
+                'Passing an instance of "%s" as the first argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                FactoryInterface::class,
+                StateMachineInterface::class,
+            ),
+        );
     }
 
     public function onPreExecute(Context $context): void
@@ -84,13 +93,19 @@ final class UpdatePaymentStateExtension implements ExtensionInterface
 
     private function updatePaymentState(PaymentInterface $payment, string $nextState): void
     {
-        $stateMachine = $this->factory->get($payment, PaymentTransitions::GRAPH);
+        $stateMachine = $this->getStateMachine();
 
-        /** @var StateMachineInterface $stateMachine */
-        Assert::isInstanceOf($stateMachine, StateMachineInterface::class);
-
-        if (null !== $transition = $stateMachine->getTransitionToState($nextState)) {
-            $stateMachine->apply($transition);
+        if (null !== $transition = $stateMachine->getTransitionToState($payment, PaymentTransitions::GRAPH, $nextState)) {
+            $stateMachine->apply($payment, PaymentTransitions::GRAPH, $transition);
         }
+    }
+
+    private function getStateMachine(): StateMachineInterface
+    {
+        if ($this->factory instanceof FactoryInterface) {
+            return new WinzouStateMachineAdapter($this->factory);
+        }
+
+        return $this->factory;
     }
 }
