@@ -15,11 +15,16 @@ namespace Sylius\Bundle\PromotionBundle\Validator;
 
 use Sylius\Bundle\PromotionBundle\Validator\CatalogPromotionAction\ActionValidatorInterface;
 use Sylius\Bundle\PromotionBundle\Validator\Constraints\CatalogPromotionAction;
-use Sylius\Component\Promotion\Model\CatalogPromotionActionInterface;
 use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintValidator;
 use Webmozart\Assert\Assert;
 
+trigger_deprecation(
+    'sylius/promotion-bundle',
+    '1.13',
+    'The "%s" class is deprecated and will be removed in Sylius 2.0, use the usual symfony logic for validation.',
+    CatalogPromotionActionValidator::class,
+);
 final class CatalogPromotionActionValidator extends ConstraintValidator
 {
     private array $actionValidators;
@@ -33,13 +38,6 @@ final class CatalogPromotionActionValidator extends ConstraintValidator
     {
         /** @var CatalogPromotionAction $constraint */
         Assert::isInstanceOf($constraint, CatalogPromotionAction::class);
-
-        /** @var CatalogPromotionActionInterface $value */
-        if (!in_array($value->getType(), $this->actionTypes, true)) {
-            $this->context->buildViolation($constraint->invalidType)->atPath('type')->addViolation();
-
-            return;
-        }
 
         $type = $value->getType();
         if (!array_key_exists($type, $this->actionValidators)) {
