@@ -103,11 +103,14 @@ interface OrderRepositoryInterface extends BaseOrderRepositoryInterface
     public function findCartByTokenValueAndChannel(string $tokenValue, ChannelInterface $channel): ?BaseOrderInterface;
 
     /**
-     * @return array<array{total: int, year: int, month: int}>
+     * @param array<string, string> $groupBy
+     *
+     * @return array<array{total: int, year: int, month: int, day: int}>
      */
-    public function getTotalPaidSalesForChannelInPeriodGroupedByYearAndMonth(
+    public function getGroupedTotalPaidSalesForChannelInPeriod(
         ChannelInterface $channel,
         \DateTimeInterface $startDate,
         \DateTimeInterface $endDate,
+        array $groupBy,
     ): array;
 }
