@@ -14,6 +14,8 @@ declare(strict_types=1);
 namespace Sylius\Component\Core\StateResolver;
 
 use SM\Factory\FactoryInterface;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
+use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
 use Sylius\Component\Core\Checker\OrderPaymentMethodSelectionRequirementCheckerInterface;
 use Sylius\Component\Core\Checker\OrderShippingMethodSelectionRequirementCheckerInterface;
 use Sylius\Component\Core\Model\OrderInterface;
@@ -25,29 +27,47 @@ use Webmozart\Assert\Assert;
 final class CheckoutStateResolver implements StateResolverInterface
 {
     public function __construct(
-        private FactoryInterface $stateMachineFactory,
+        private FactoryInterface|StateMachineInterface $stateMachineFactory,
         private OrderPaymentMethodSelectionRequirementCheckerInterface $orderPaymentMethodSelectionRequirementChecker,
         private OrderShippingMethodSelectionRequirementCheckerInterface $orderShippingMethodSelectionRequirementChecker,
     ) {
+        trigger_deprecation(
+            'sylius/core',
+            '1.13',
+            sprintf(
+                'Passing an instance of "%s" as the first argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                FactoryInterface::class,
+                StateMachineInterface::class,
+            ),
+        );
     }
 
     public function resolve(BaseOrderInterface $order): void
     {
         Assert::isInstanceOf($order, OrderInterface::class);
-        $stateMachine = $this->stateMachineFactory->get($order, OrderCheckoutTransitions::GRAPH);
+        $stateMachine = $this->getStateMachine();
 
         if (
             !$this->orderShippingMethodSelectionRequirementChecker->isShippingMethodSelectionRequired($order) &&
-            $stateMachine->can(OrderCheckoutTransitions::TRANSITION_SKIP_SHIPPING)
+            $stateMachine->can($order, OrderCheckoutTransitions::GRAPH, OrderCheckoutTransitions::TRANSITION_SKIP_SHIPPING)
         ) {
-            $stateMachine->apply(OrderCheckoutTransitions::TRANSITION_SKIP_SHIPPING);
+            $stateMachine->apply($order, OrderCheckoutTransitions::GRAPH, OrderCheckoutTransitions::TRANSITION_SKIP_SHIPPING);
         }
 
         if (
             !$this->orderPaymentMethodSelectionRequirementChecker->isPaymentMethodSelectionRequired($order) &&
-            $stateMachine->can(OrderCheckoutTransitions::TRANSITION_SKIP_PAYMENT)
+            $stateMachine->can($order, OrderCheckoutTransitions::GRAPH, OrderCheckoutTransitions::TRANSITION_SKIP_PAYMENT)
         ) {
-            $stateMachine->apply(OrderCheckoutTransitions::TRANSITION_SKIP_PAYMENT);
+            $stateMachine->apply($order, OrderCheckoutTransitions::GRAPH, OrderCheckoutTransitions::TRANSITION_SKIP_PAYMENT);
         }
+    }
+
+    private function getStateMachine(): StateMachineInterface
+    {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            return new WinzouStateMachineAdapter($this->stateMachineFactory);
+        }
+
+        return $this->stateMachineFactory;
     }
 }
