@@ -16,7 +16,10 @@ namespace Sylius\Component\Core\Updater;
 use Doctrine\Persistence\ObjectManager;
 use Psr\Log\LoggerInterface;
 use SM\Factory\Factory;
-use SM\SMException;
+use SM\Factory\FactoryInterface;
+use Sylius\Abstraction\StateMachine\Exception\StateMachineExecutionException;
+use Sylius\Abstraction\StateMachine\StateMachineInterface;
+use Sylius\Abstraction\StateMachine\WinzouStateMachineAdapter;
 use Sylius\Component\Core\Repository\OrderRepositoryInterface;
 use Sylius\Component\Order\Model\OrderInterface;
 use Sylius\Component\Order\OrderTransitions;
@@ -25,12 +28,25 @@ final class UnpaidOrdersStateUpdater implements UnpaidOrdersStateUpdaterInterfac
 {
     public function __construct(
         private OrderRepositoryInterface $orderRepository,
-        private Factory $stateMachineFactory,
+        private Factory|StateMachineInterface $stateMachineFactory,
         private string $expirationPeriod,
         private LoggerInterface $logger,
         private ObjectManager $orderManager,
         private int $batchSize = 100,
     ) {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            trigger_deprecation(
+                'sylius/core',
+                '1.13',
+                sprintf(
+                    'Passing an instance of "%s" as the second argument is deprecated. It will accept only instances of "%s" in Sylius 2.0.',
+                    FactoryInterface::class,
+                    StateMachineInterface::class,
+                ),
+            );
+        }
+
+        $this->logger = $logger;
     }
 
     public function cancel(): void
@@ -56,13 +72,22 @@ final class UnpaidOrdersStateUpdater implements UnpaidOrdersStateUpdaterInterfac
     private function cancelOrder(OrderInterface $expiredUnpaidOrder): void
     {
         try {
-            $stateMachine = $this->stateMachineFactory->get($expiredUnpaidOrder, OrderTransitions::GRAPH);
-            $stateMachine->apply(OrderTransitions::TRANSITION_CANCEL);
-        } catch (SMException $e) {
+            $stateMachine = $this->getStateMachine();
+            $stateMachine->apply($expiredUnpaidOrder, OrderTransitions::GRAPH, OrderTransitions::TRANSITION_CANCEL);
+        } catch (StateMachineExecutionException $e) {
             $this->logger->error(
                 sprintf('An error occurred while cancelling unpaid order #%s', $expiredUnpaidOrder->getId()),
                 ['exception' => $e, 'message' => $e->getMessage()],
             );
         }
+    }
+
+    private function getStateMachine(): StateMachineInterface
+    {
+        if ($this->stateMachineFactory instanceof FactoryInterface) {
+            return new WinzouStateMachineAdapter($this->stateMachineFactory);
+        }
+
+        return $this->stateMachineFactory;
     }
 }
